Allow users to call error_estimation successfully.
This commit adds the functionality of error estimation in clad. Users can now call clad::error_estimation on a target function to generate the error estimation code for that function. Users will have the choice to either write their own error model which will be used by clad or use the default built-in model.